split algorithms into their own files
